move prune to pkg/resourceinterpreter/defaultinterpreter
