+++ pkg/resourceinterpreter/defaultinterpreter/prune/prune.go
